Remove magic number 96 (= 65536 S2K iterations)
In the case where no Config.S2KCount was given, the S2K iteration count
is set to 65536.

This was obscured by the fact that this function returned a magic number
96 (which the comment claims is the pre-computed value for S2K 65535
iterations).

Rather than returning a magic number, allow the encode function to
calculate with the default value S2KCountDefault.